requirements.txt
# --- Web framework ---
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
requests==2.31.0

# --- Data models ---
pydantic==2.7.1

# --- Database ---
psycopg2-binary==2.9.9

# --- Machine learning / NLP ---
scikit-learn==1.3.0
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
gensim==4.3.2
nltk==3.8.1

# --- File processing ---
pdfminer.six==20221105
python-docx==1.1.0
PyPDF2==3.0.1

# --- Visualization ---
plotly==5.22.0
